made the obj-c Protocol better, made the game API calls much more efficient, made the game errors have descriptions, made the random game list item function, made a url class for API calls (games)
