+++ TechReadsPod/CkickenCoopAPI/GameAPI/DisplayProtocol.h
@@ -11,8 +11,8 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-@protocol DisplayProtocol
-- (void) todisplay;
+@protocol GamePresenterView
+- (void) updatePageContent : (NSString *) title withContent: (NSString*)  content andImageURL: (NSString*) imageUrl;
 @end
 
 NS_ASSUME_NONNULL_END
