+++ TechReadsPod/NewsAPI/NewsAPIModel/NewsErrors.swift
